Extract AutoCompletionUseCase
