+++ feature/metis/conversation/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/ui/ConversationChatListScreen.kt
@@ -134,7 +134,7 @@ internal fun ConversationChatListScreen(
         val isReplyEnabled = isReplyEnabled(conversationDataState = conversationDataState)
 
         if (conversationDataState.isSuccess) {
-            CompositionLocalProvider(LocalReplyAutoCompleteHintProvider provides viewModel) {
+            CompositionLocalProvider(LocalReplyAutoCompleteHintProvider provides viewModel.autoCompletionUseCase) {
                 MetisChatList(
                     modifier = Modifier
                         .fillMaxSize()
