+++ feature/metis/conversation/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/ui/ConversationThreadScreen.kt
@@ -84,7 +84,7 @@ internal fun ConversationThreadScreen(
                 .fillMaxSize()
                 .imePadding(),
         ) {
-            CompositionLocalProvider(LocalReplyAutoCompleteHintProvider provides viewModel) {
+            CompositionLocalProvider(LocalReplyAutoCompleteHintProvider provides viewModel.autoCompletionUseCase) {
                 MetisThreadUi(
                     modifier = Modifier
                         .fillMaxWidth()
